Simplify implementation for now
