+++ Samples/SwiftUIExample/SwiftUIExampleUITests/SwiftUIExampleUITestsLaunchTests.swift
@@ -24,11 +24,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SO
 import XCTest
 
 final class SwiftUIExampleUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
     override func setUpWithError() throws {
         continueAfterFailure = false
     }
